feat/item: Slider의 ltem 요소 컴포넌트화, 일부 에러 수정
