+++ client/src/api/api.ts
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import { Member, NewMember, LoginInfo, TV } from '../types/types';
+import { Member, NewMember, LoginInfo, ItemData } from '../types/types';
 
 const accessToken = localStorage.getItem('token');
 
@@ -33,5 +33,5 @@ export const PatchUser = (data: any) => instance.patch(`/members`, data);
 export const DeleteUser = () => instance.delete('/members');
 
 /* TV 데이터 가져오기 */
-export const GetTVData = (): Promise<TV[]> =>
+export const GetTVData = (): Promise<ItemData[]> =>
   instance.get('/TV').then((res) => res.data);
