+++ client/src/components/Banner.tsx
@@ -1,10 +1,9 @@
 import styled from 'styled-components';
-import BannerImage1 from '../assets/이번생도잘부탁해.webp';
 
-const Banner = () => {
+const Banner = ({ image }: { image: string }) => {
   return (
     <S_Wrapper>
-      <BannerImage src={BannerImage1} alt='이번생도잘부탁해'></BannerImage>
+      <BannerImage src={image} alt={image}></BannerImage>
       <BlackLinear></BlackLinear>
     </S_Wrapper>
   )
@@ -20,11 +19,8 @@ const S_Wrapper = styled.div`
 const BannerImage = styled.img`
   display: block;
   object-fit: cover;
-  /* left: 0px;
-  right: 0px; */
   width: 100%;
   height: 641px;
-  /* height: 100%; */
 `
 
 const BlackLinear = styled.div`
